+++ HaishinKit/Sources/Extension/VTCompressionSession+Extension.swift
@@ -9,17 +9,24 @@ extension VTCompressionSession {
 
 extension VTCompressionSession: VTSessionConvertible {
     @inline(__always)
-    func convert(_ sampleBuffer: CMSampleBuffer, continuation: AsyncStream<CMSampleBuffer>.Continuation?) throws {
+    func convert(
+        _ sampleBuffer: CMSampleBuffer,
+        forceKeyFrame: Bool,
+        continuation: AsyncStream<CMSampleBuffer>.Continuation?
+    ) throws {
         guard let imageBuffer = sampleBuffer.imageBuffer else {
             return
         }
         var flags: VTEncodeInfoFlags = []
+        let frameProperties = forceKeyFrame ? [
+            VTSessionOptionKey.forceKeyFrame.CFString: kCFBooleanTrue as Any
+        ] as CFDictionary : nil
         let status = VTCompressionSessionEncodeFrame(
             self,
             imageBuffer: imageBuffer,
             presentationTimeStamp: sampleBuffer.presentationTimeStamp,
             duration: sampleBuffer.duration,
-            frameProperties: nil,
+            frameProperties: frameProperties,
             infoFlagsOut: &flags,
             outputHandler: { _, _, sampleBuffer in
                 if let sampleBuffer {