+++ SRTHaishinKit/Sources/SRT/SRTStream.swift
@@ -153,7 +153,10 @@ public actor SRTStream {
 
     private func startMixerInputConsumers() {
         let (audioStream, audioContinuation) = AsyncStream.makeStream(of: (AVAudioPCMBuffer, AVAudioTime).self)
-        let (videoStream, videoContinuation) = AsyncStream.makeStream(of: CMSampleBuffer.self)
+        let (videoStream, videoContinuation) = AsyncStream.makeStream(
+            of: CMSampleBuffer.self,
+            bufferingPolicy: .bufferingNewest(outgoing.videoInputBufferCounts)
+        )
         mixerAudioContinuation = audioContinuation
         mixerVideoContinuation = videoContinuation
         Task {